Airflow_Google_Sheets/requirements.txt
pymssql
pandas
sqlalchemy
gspread
psycopg2-binary
openpyxl